Added new Location model for rat sightings
