+++ Ratastic/app/src/main/java/edu/gatech/teamraid/ratastic/Model/Location.java
@@ -1,0 +1,86 @@
+package edu.gatech.teamraid.ratastic.Model;
+
+import java.util.ArrayList;
+
+/**
+ * Created by colby on 10/18/17.
+ */
+
+public class Location {
+
+    private String locationType;
+    private String incidentZip;
+    private String incidentAddress;
+    private String city;
+    private String borough;
+    private float lat;
+    private float lng;
+
+    public Location(String locationType, String incidentZip,
+                       String incidentAddress, String city, String borough, float lat, float lng) {
+        this.locationType = locationType;
+        this.incidentZip = incidentZip;
+        this.incidentAddress = incidentAddress;
+        this.city = city;
+        this.borough = borough;
+        this.lat = lat;
+        this.lng = lng;
+    }
+
+    public String getLocationType() {
+        return locationType;
+    }
+
+    public void setLocationType(String locationType) {
+        this.locationType = locationType;
+    }
+
+    public String getIncidentZip() {
+        return incidentZip;
+    }
+
+    public void setIncidentZip(String incidentZip) {
+        this.incidentZip = incidentZip;
+    }
+
+    public String getIncidentAddress() {
+        return incidentAddress;
+    }
+
+    public void setIncidentAddress(String incidentAddress) {
+        this.incidentAddress = incidentAddress;
+    }
+
+    public String getCity() {
+        return city;
+    }
+
+    public void setCity(String city) {
+        this.city = city;
+    }
+
+    public String getBorough() {
+        return borough;
+    }
+
+    public void setBorough(String borough) {
+        this.borough = borough;
+    }
+
+    public float getLat() {
+        return lat;
+    }
+
+    public void setLat(float lat) {
+        this.lat = lat;
+    }
+
+    public float getLng() {
+        return lng;
+    }
+
+    public void setLng(float lng) {
+        this.lng = lng;
+    }
+
+}
