+++ Ratastic/app/src/main/java/edu/gatech/teamraid/ratastic/Model/RatSighting.java
@@ -10,27 +10,16 @@ public class RatSighting {
 
     private String UID;
     private String createdDate;
-    private String locationType;
-    private String incidentZip;
-    private String incidentAddress;
-    private String city;
-    private String borough;
-    private float lat;
-    private float lng;
+
+    private Location location;
 
     public static ArrayList<RatSighting> ratSightingArray = new ArrayList<>();
 
-    public RatSighting(String UID, String createdDate, String locationType, String incidentZip,
-                       String incidentAddress, String city, String borough, float lat, float lng) {
+    public RatSighting(String UID, String createdDate, Location location) {
         this.UID = UID;
         this.createdDate = createdDate;
-        this.locationType = locationType;
-        this.incidentZip = incidentZip;
-        this.incidentAddress = incidentAddress;
-        this.city = city;
-        this.borough = borough;
-        this.lat = lat;
-        this.lng = lng;
+        this.location = location;
+
     }
 
     public String getUID() {
@@ -49,63 +38,13 @@ public class RatSighting {
         this.createdDate = createdDate;
     }
 
-    public String getLocationType() {
-        return locationType;
+    public void setLocation(Location location) {
+        this.location = location;
     }
 
-    public void setLocationType(String locationType) {
-        this.locationType = locationType;
-    }
-
-    public String getIncidentZip() {
-        return incidentZip;
-    }
-
-    public void setIncidentZip(String incidentZip) {
-        this.incidentZip = incidentZip;
-    }
-
-    public String getIncidentAddress() {
-        return incidentAddress;
-    }
-
-    public void setIncidentAddress(String incidentAddress) {
-        this.incidentAddress = incidentAddress;
-    }
-
-    public String getCity() {
-        return city;
-    }
-
-    public void setCity(String city) {
-        this.city = city;
-    }
-
-    public String getBorough() {
-        return borough;
-    }
-
-    public void setBorough(String borough) {
-        this.borough = borough;
-    }
-
-    public float getLat() {
-        return lat;
-    }
-
-    public void setLat(float lat) {
-        this.lat = lat;
-    }
-
-    public float getLng() {
-        return lng;
-    }
-
-    public void setLng(float lng) {
-        this.lng = lng;
-    }
+    public Location getLocation() { return location; }
 
     public String toString() {
-        return "Sighting in " + city;
+        return "Sighting in " + location.getCity();
     }
 }
